Add TShareCommand trait
